+++ Food.java
@@ -1,6 +1,6 @@
 public class Food extends MenuItem {
 
-    public Food(String name, int cost) {
+    public Food(String name, float cost) {
         super(name, cost);
     }
